+++ sorter/config.py
@@ -1,0 +1,2 @@
+
+IMAGE_EXTENSIONS = {".jpg", ".jpeg", ".gif", ".png", ".webp", ".heif", ".heic", ".tiff", ".bmp", ".raw", ".svg"}