+++ CHANGELOG.md
@@ -6,3 +6,9 @@ All commits will be recorded here.
 ## [1.0.0] Initial Commit
 
 - Created `main.py`, `CHANGELOG.md`, `README.md` and `requirements.txt`
+
+---
+
+## [1.0.1] `file_browser` and `read_source` functions in utils.py
+
+- Added `file_browser` and `read_source` functions to utils.py